+++ src/examples/function/editPackage.tsx
@@ -2,9 +2,9 @@ const delay = (ms: number) => {
   return new Promise((resolve) => setTimeout(resolve, ms))
 }
 let nieco =
-  '<svg xmlns="http://www.w3.org/2000/svg"><rect stroke="#ff0000" opacity="1" stroke-width="15" stroke-linecap="round" fill="none" rx="50%" x="148" y="238" width="200" height="300"></rect></svg>'
+  '<svg xmlns="http://www.w3.org/2000/svg" width="362" height="549"><rect stroke="#ff0000" opacity="1" stroke-width="16" stroke-linecap="round" fill="none" rx="50%" x="101" y="189" width="167" height="205"></rect></svg>'
 let nieco1 =
-  '<svg xmlns="http://www.w3.org/2000/svg"><rect stroke="#ff0000" opacity="1" stroke-width="15" stroke-linecap="round" fill="none" rx="50%" x="148" y="238" width="220" height="300"></rect></svg>'
+  '<svg xmlns="http://www.w3.org/2000/svg" width="362" height="549"><rect stroke="#ff0000" opacity="1" stroke-width="16" stroke-linecap="round" fill="none" rx="50%" x="101" y="189" width="167" height="205"></rect></svg>'
 
 export const saveLayerFunc = async (
   svg: HTMLElement,